% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fixed_levels}
\alias{fixed_levels}
\title{Fix a level to the last in sequence for factor subject}
\usage{
fixed_levels(s, f)
}
\arguments{
\item{s}{Vector to be  factorized}

\item{f}{The target level to be fixed in the last}
}
\value{
Factorized variable
}
\description{
Fix a level to the last in sequence for factor subject
}
\examples{
fixed_levels(s=c("a","b","d","e"),f="b")

}
