% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_boxplot.R
\name{gemini2_boxplot}
\alias{gemini2_boxplot}
\title{GEMINI2 boxplot visualization function}
\usage{
gemini2_boxplot(opt, gene_pair, cell_line, nc_gene, show_inference = T)
}
\arguments{
\item{opt}{The list of inference results from \code{\link[gemini2]{gemini2_inference}} (list)}

\item{gene_pair}{The gene pair whose LFC data is used for visualization  (character)}

\item{cell_line}{The cell line name for extracting LFC data (character)}

\item{nc_gene}{A character vector naming the genes to use as a negative control,
to be paired with each individual g and h in a gene pair. (character)}

\item{show_inference}{A logical indicating whether to show the
inferred individual or combined values for each gene/gene pair (logical)}
}
\value{
A ggplot2 object
}
\description{
A function to visualize the results of GEMINI over the raw data.
}
\details{
Raw LFC data is plotted for each gene combination (\code{g}-\code{nc_gene}, \code{h}-\code{nc_gene}, \code{g}-\code{h}) in a standard boxplot.
Horizontal green line segments are plotted over the box plots indicating the individual gene effects or
the inferred total effect of a particular gene combination. Each guide
pair can be colored based on the inferred sample independent effects
\eqn{g_i}, \eqn{h_j}, and \eqn{g_i,h_j}. Additionally, colors and shapes
can be used to distinguish unique guides targeting gene g and h, respectively.
}
\examples{
p_box <- gemini2_boxplot(opt=infered_D,gene_pair="HDAC1/HDAC2",cell_line="MELJUSO_SKIN",nc_gene="AAVS1",show_inference=T)

}
