% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_confusion.R
\name{gemini2_confusion}
\alias{gemini2_confusion}
\title{GEMINI2 confusion matrix generating function}
\usage{
gemini2_confusion(
  score,
  pos_ctrl,
  neg_ctrl,
  cut_seq = NULL,
  hit_cl = 1,
  cl = NULL,
  direction = "geq"
)
}
\arguments{
\item{score}{The list of scoring results from \code{\link[gemini2]{gemini2_score}} (list)}

\item{pos_ctrl}{A character vector of gene pairs serving as positive controls (character)}

\item{neg_ctrl}{A character vector of gene pairs serving as negative controls (character)}

\item{cut_seq}{A series of discrimination thresholds (numeric)}

\item{hit_cl}{The number of cell lines where the score of an interaction at least needs to fulfill discrimination threshold
in order to be counted as true positive  (numeric)}

\item{cl}{A character specifying a cell line to calculate confusion matrix (character)}
}
\value{
A confusion matrix
}
\description{
The function calculates the confusion matrix for llustrating the diagnostic ability of a binary classifier system as its discrimination threshold is varied.
}
\details{
Between parameters \code{hit_cl} and \code{cl}, only 1 need to be specified by the user
}
\examples{
confusion_m <- gemini2_confusion(score=scores$sensitive_lethality,
pos_ctrl = c("AURKB;PLK4","BAP1;UCHL5","BCL2L1;BCL2L2","BCL2L1;MCL1"),
neg_ctrl = c("CCL13;CCL8","CDY1;CDY1B","CDY1;CDY2A"),cut_seq=seq(10, -10, by = -0.01),hit_cl=1,cl=NULL)

}
