% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_costplot.R
\name{gemini2_costplot}
\alias{gemini2_costplot}
\title{GEMINI2 cost plotting function}
\usage{
gemini2_costplot(costs)
}
\arguments{
\item{input}{A series of costs (numeric)}
}
\value{
A ggplot2 object
}
\description{
The function visualizes a series of costs with dot plot
}
\examples{
p <- gemini2_costplot(infered_D$cost)

}
