% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_inference.R
\name{gemini2_inference}
\alias{gemini2_inference}
\title{GEMINI2 Inference function}
\usage{
gemini2_inference(
  input,
  method = "min",
  gamma = NULL,
  mu = 0.1,
  nc_zero = T,
  cores = NULL,
  n_iterations = 21,
  verbose = F
)
}
\arguments{
\item{input}{Initialized object from \code{\link[gemini2]{gemini2_initialization}} (list)}

\item{method}{Link function for modeling combined guide effect, "min" recommended for asymmetric combinatory screen and "prod" recommended for symmetric one. (character)}

\item{gamma}{2xthe penalized factor set for double gene effect. When NULL is specificed, the function calculates it as 0.05 x (Mean guide pairs per gene pair) x quantile 50\% of the initialized inverse of variance (numeric)}

\item{mu}{2*the penalized factor set for single gene effect. Default is 0.1 (numeric)}

\item{nc_zero}{Logical flag for whether double gene effects for pairs with negative control and single gene effects for negative control genes are set to zero or not. Default is TRUE (logical)}

\item{cores}{The cores to be used for code parallelization. If not specified, only 2 codes will be spared out when parallelization. (numeric)}

\item{n_iterations}{Number of iteractions +1 to run for coordinate ascent optimization (numeric)}

\item{verbose}{Logical flag for whether to print all message out (logical)}
}
\value{
A list of optimized guide/gene effect, inverse of variance, the calculated LFC, corresponding weights and the series of costs for optimization
}
\description{
The function uses coordinate ascent method to infer a series of gene/guide effects.
}
\details{
Link functions "max" and "sqrt" for also available for \code{method}
}
\examples{
infered_D <- gemini2_inference(input=init_D,method="min",gamma=NULL,mu=0.1,nc_zero=T,cores=NULL,n_iterations=21,verbose=F)

}
