% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_initilization.R
\name{gemini2_initilization}
\alias{gemini2_initilization}
\title{GEMINI2 Initilization function}
\usage{
gemini2_initilization(
  counts,
  replicate.map,
  guide.annot,
  nc_gene,
  a = 0.01,
  cores = NULL,
  gene_join = "/"
)
}
\arguments{
\item{counts}{Raw guide pair count data obtained from NGS reads (matrix)}

\item{replicate.map}{Mapping between replicate name and corresponding cell line name (data.frame)}

\item{guide.annot}{Mapping between guide pair string,the row names of count matrix,to genes in two sides (data.frame)}

\item{nc_gene}{A vector or single character string for negative control genes (character)}

\item{a}{The noise to be added/subtracted from extreme values of guide effect. Default is 0.01 (numeric)}

\item{cores}{The cores to be used for code parallelization. If not specified, only 2 codes will be spared out when parallelization. (number)}

\item{gene_join}{The separator to join the gene pairs, default "/" (character)}
}
\value{
a list of initialized variables including guide effect \code{X}, single gene effect \code{Y}, double gene effecr \code{S}
and replicate variance; the computed LFC, processed mapping data, the input a as well as negative control gene(s).
}
\description{
This function processes from raw screening counts to initialize
a list of guide effect, single/double effect, sample variance for inference
}
\details{
counts should be a matrix which rownames are guide pairs from the first column of \code{guide.annot} and colnames are sample replicate names
replicate.map should be a data frame with three columns in the order:
\itemize{
\item 1st: replicate name, the names of sample replicates that has been screened, contain \code{colnames} of counts matrix
\item 2nd: sample name, the cell line names corresponding to replicate names in the first column
\item 3rd: label for time point, a factor (Two levels: "ETP","LTP") column specifying which replicates were screened in early time point.
}
guide.annot should be a data.frame with three columns in the order:
\itemize{
\item 1st: guide pairs, the pattern used to join two guides here should be the same for rownames of counts
\item 2nd: gene name mapping to the first/left guide in the 1st column.
\item 3rd: gene name mapping to the second/right guide in the 1st column.
}
}
\examples{
init_D <- gemini2_initilization(counts=COUNTS,replicate.map = repMap,guide.annot=guideAnnotation,a=0.05,nc_gene="AAVS1",cores=NULL,gene_join = "/")

}
