% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_PR.R
\name{gemini2_PR}
\alias{gemini2_PR}
\title{GEMINI2 Precision-Recall curve plotting function}
\usage{
gemini2_PR(confusion, label = NULL, single = T)
}
\arguments{
\item{confusion}{A confusion table obtained from \code{\link[gemini2]{gemini2_confusion}} (data.frame)}

\item{label}{The column name for specifying group variable}
}
\value{
A ggplot2 object
}
\description{
The function plots Precision-Recall curve
}
\examples{
PR_curve <- gemini2_PR(confusion = confusion_m)

}
