% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_roc.R
\name{gemini2_roc}
\alias{gemini2_roc}
\title{GEMINI2 ROC curve plotting function}
\usage{
gemini2_roc(confusion, label = NULL, single = T)
}
\arguments{
\item{confusion}{A confusion table obtained from \code{\link[gemini2]{gemini2_confusion}} (data.frame)}

\item{label}{The column name for specifying group variable}
}
\value{
A ggplot2 object
}
\description{
The function plots ROC curve
}
\examples{
roc_curve <- gemini2_roc(confusion = confusion_m)

}
