% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini2_score.R
\name{gemini2_score}
\alias{gemini2_score}
\title{GEMINI2 genetic interaction scoring function}
\usage{
gemini2_score(
  opt,
  pc_threshold = -1,
  pc_gene = NULL,
  pc_weight = 0.5,
  nc_pairs = NULL
)
}
\arguments{
\item{opt}{The list of inference results from \code{\link[gemini2]{gemini2_inference}} (list)}

\item{pc_threshold}{A numeric value to indicate the LFC corresponding to a positive control,
if no pc_gene is specified. Default is -1 (numeric)}

\item{pc_gene}{A character vector of any length naming genes to use as positive controls (character)}

\item{pc_weight}{A weighting applied to the positive control
(\code{pc_gene}) to filter genes whose individual phenotype is
more lethal than \eqn{pc_weight*y(pc_gene)}. Default is 0.5 (numeric)}

\item{nc_pairs}{A set of non-interacting gene pairs to define statistical significance. (character)}
}
\value{
An object of class gemini.score containing score values for
strong interactions and sensitive lethality and recovery, and if \code{nc_pairs} is specified, statistical significance for each scoring type.
}
\description{
The function scores genetic interactions based on inference results
}
\examples{
scores <- gemini2_score(opt =infered_D, pc_gene = "EEF2")

}
