% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{matrix_conv}
\alias{matrix_conv}
\title{Matrix conversion}
\usage{
matrix_conv(m)
}
\arguments{
\item{m}{Data.frame whose 1st column will serve as rownames}
}
\value{
Converted matrix
}
\description{
Matrix conversion
}
\examples{
matrix_conv(init_D$Y)

}
