% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simple_auc}
\alias{simple_auc}
\title{AUC calculation with series of axis values}
\usage{
simple_auc(y, x)
}
\arguments{
\item{y}{Ordered values on Y axis}

\item{x}{Ordered values on X axis}
}
\value{
Calculated simple auc
}
\description{
AUC calculation with series of axis values
}
\examples{
simple_auc(y = seq(1:20),x=seq(1:20))

}
